# docs/api/swagger.yaml
basePath: /api/v1
definitions:
...
  github_com_opentrusty_opentrusty_internal_tenant.Tenant:
    properties:
      created_at:
        type: string
      id:
        type: string
      name:
        type: string
      status:
        type: string
      updated_at:
        type: string
    type: object
  github_com_opentrusty_opentrusty_internal_tenant.TenantUserRole:
    properties:
      granted_at:
        type: string
      granted_by:
        type: string
      id:
        type: string
      role:
        type: string
      tenant_id:
        type: string
      user_id:
        type: string
    type: object
  internal_transport_http.AssignRoleRequest:
    properties:
      role:
        example: member
        type: string
    required:
    - role
    type: object
...
  internal_transport_http.CreateTenantRequest:
    properties:
      id:
        example: tenant-1
        type: string
      name:
        example: My Corporation
        type: string
    required:
    - id
    - name
    type: object
...
  internal_transport_http.ProvisionUserRequest:
    properties:
      email:
        example: user@example.com
        type: string
      family_name:
        example: Doe
        type: string
      given_name:
        example: John
        type: string
      password:
        example: secret123
        type: string
      role:
        example: admin
        type: string
    required:
    - email
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: Production-grade Identity Provider
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: OpenTrusty API
  version: 1.0.0
paths:
...
  /tenants:
    post:
      consumes:
      - application/json
      description: Create a new tenant
      parameters:
      - description: Tenant Data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_transport_http.CreateTenantRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_opentrusty_opentrusty_internal_tenant.Tenant'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - CookieAuth: []
      summary: Create Tenant
      tags:
      - Tenant
  /tenants/{tenantID}/users:
    get:
      description: List all users and their roles in a tenant
      parameters:
      - description: Tenant ID
        in: path
        name: tenantID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/github_com_opentrusty_opentrusty_internal_tenant.TenantUserRole'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - CookieAuth: []
      summary: List Tenant Users
      tags:
      - Tenant
    post:
      consumes:
      - application/json
      description: Create a user and assign a role within a tenant
      parameters:
      - description: Tenant ID
        in: path
        name: tenantID
        required: true
        type: string
      - description: User Data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_transport_http.ProvisionUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - CookieAuth: []
      summary: Provision Tenant User
      tags:
      - Tenant
  /tenants/{tenantID}/users/{userID}/roles:
    post:
      consumes:
      - application/json
      description: Assign a role to a user within a tenant
      parameters:
      - description: Tenant ID
        in: path
        name: tenantID
        required: true
        type: string
      - description: User ID
        in: path
        name: userID
        required: true
        type: string
      - description: Role Data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_transport_http.AssignRoleRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - CookieAuth: []
      summary: Assign Role
      tags:
      - Tenant
  /tenants/{tenantID}/users/{userID}/roles/{role}:
    delete:
      description: Revoke a role from a user within a tenant
      parameters:
      - description: Tenant ID
        in: path
        name: tenantID
        required: true
        type: string
      - description: User ID
        in: path
        name: userID
        required: true
        type: string
      - description: Role
        in: path
        name: role
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - CookieAuth: []
      summary: Revoke Role
      tags:
      - Tenant
...
securityDefinitions:
  CookieAuth:
    in: cookie
    name: session_id
    type: apiKey
swagger: "2.0"